+++ test/persistence/redis_spec.js
@@ -6,6 +6,8 @@ var redis = require("redis");
 
 describe("mosca.persistence.Redis", function() {
 
+  this.timeout(5000);
+
   var opts = {
     ttl: {
       subscriptions: 1000,
@@ -179,6 +181,48 @@ describe("mosca.persistence.Redis", function() {
               redisClient.llen("packets:" + client.id, function(err, length) {
                 expect(length).to.eql(1);
                 done();
+              });
+            }, 50);
+          });
+        });
+      });
+    });
+
+  });
+
+
+  describe("clustered.environment", function(){
+
+    it("should forward each packet once after client reconnects", function(done) {
+      var client = {
+        id: "cluster client id - 42",
+        clean: false,
+        subscriptions: {
+          "hello/#": {
+            qos: 1
+          }
+        }
+      };
+
+      var packet = {
+        topic: "hello/42",
+        qos: 0,
+        payload: new Buffer("world"),
+        messageId: "42"
+      };
+
+      var that = this;
+      that.secondInstance = new Redis(opts, function() {
+        that.instance.storeSubscriptions(client, function() {
+          // simulate client reconnect since storeSubscriptions is called on disconnect
+          // no matter client connects to instance or secondInstance
+          that.secondInstance.storeSubscriptions(client, function () {
+            setTimeout(function () {
+              that.secondInstance.storeOfflinePacket(packet, function () {
+                that.instance.streamOfflinePackets(client, function (err, p) {
+                  expect(p).to.eql(packet);
+                  done(); // should be called once
+                });
               });
             }, 50);
           });