+++ HISTORY.md
@@ -1,6 +1,15 @@
 History
 =======
 
+## 0.29.0
+
+* Effectively support offline packets expiration #235 #236 by @mazhack.
+
+## 0.28.2
+
+* Do not store qos 0 packets for offline delivery #229 #230 by @mazhack.
+* Added session present flag on CONNACK #231 by @mazhack).
+
 ## 0.28.1
 
 * Fixed "stats should not be published by default in embedded mode"